+++ flink-connectors/pom.xml
@@ -57,6 +57,7 @@ under the License.
 		<module>flink-connector-gcp-pubsub</module>
 		<module>flink-connector-kinesis</module>
 		<module>flink-sql-connector-elasticsearch7</module>
+		<module>flink-connector-base</module>
 	</modules>
 
 	<!-- override these root dependencies as 'provided', so they don't end up
